Move MS2-specific SequencesTableInfo customizations into MS2Schema & MS2Controller, and then move it, ProteinUserSchema, CustomAnnotationSchema, and a bunch of other user tables into the Protein module.
