+++ ms2/src/org/labkey/ms2/protein/ProteinController.java
@@ -39,6 +39,7 @@ import org.labkey.api.protein.CustomAnnotationSet;
 import org.labkey.api.protein.CustomAnnotationSetManager;
 import org.labkey.api.protein.CustomAnnotationType;
 import org.labkey.api.protein.ProteinSchema;
+import org.labkey.api.protein.query.CustomAnnotationSchema;
 import org.labkey.api.query.QuerySettings;
 import org.labkey.api.query.QueryView;
 import org.labkey.api.query.UserSchema;
@@ -60,7 +61,6 @@ import org.labkey.api.view.JspView;
 import org.labkey.api.view.NavTree;
 import org.labkey.api.view.NotFoundException;
 import org.labkey.api.view.VBox;
-import org.labkey.ms2.protein.query.CustomAnnotationSchema;
 import org.springframework.validation.BindException;
 import org.springframework.validation.Errors;
 import org.springframework.validation.ObjectError;
